+++ PyHotKey/__init__.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 #
-# Copyright (C) 2019-2023 Xpp521
+# Copyright (C) 2019-2024 Xpp521
 #
 # This program is free software: you can redistribute it and/or modify it under
 # the terms of the GNU Lesser General Public License as published by the Free
@@ -17,6 +17,9 @@
 """
                                 PyHotKey
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-                A cross-platform hotkey module for Python.
+                A cross-platform keyboard module for Python.
 """
-from ._manager import Key, keyboard_manager
+from ._keys import Key
+from ._keyboard import keyboard
+
+__all__ = ['Key', 'keyboard']