+++ PyHotKey/_info.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 #
-# Copyright (C) 2019-2023 Xpp521
+# Copyright (C) 2019-2024 Xpp521
 #
 # This program is free software: you can redistribute it and/or modify it under
 # the terms of the GNU Lesser General Public License as published by the Free
@@ -17,4 +17,4 @@
 NAME = 'PyHotKey'
 AUTHOR = 'Xpp'
 AUTHOR_EMAIL = 'xpp233@foxmail.com'
-VERSION = '1.5.0'
+VERSION = '1.5.2'